+++ src/filters/shaders/saturation.ts
@@ -1,0 +1,15 @@
+export const fragmentSource = `
+  precision highp float;
+  uniform sampler2D uTexture;
+  uniform float uSaturation;
+  varying vec2 vTexCoord;
+  void main() {
+    vec4 color = texture2D(uTexture, vTexCoord);
+    float rgMax = max(color.r, color.g);
+    float rgbMax = max(rgMax, color.b);
+    color.r += rgbMax != color.r ? (rgbMax - color.r) * uSaturation : 0.00;
+    color.g += rgbMax != color.g ? (rgbMax - color.g) * uSaturation : 0.00;
+    color.b += rgbMax != color.b ? (rgbMax - color.b) * uSaturation : 0.00;
+    gl_FragColor = color;
+  }
+`;