+++ test/visual/control_rendering.js
@@ -27,7 +27,7 @@
   fabric.config.configure({
     enableGLFiltering: false
   });
-  fabric.Object.prototype.objectCaching = true;
+  fabric.Object.ownDefaults.objectCaching = true;
   var visualTestLoop;
   if (isNode()) {
     visualTestLoop = global.visualTestLoop;